Rebuild the commandline UI with Click
This splits the "bot" (CI) usage and the "release" usage into different
modules, and adds another module for the CLI, splitting it out to make it
easier to reuse the pieces.

The logging was updated to use `contextvars` in modern structlog, and the CLI
can also ask about the milestone to use.