milestone Issue charts and lists
This adds a new data class for Issues and splits out the Issue parsing logic
from the issue printing/changelog code.

I kept the Loss-first logic rather than changing much of the behaviour of the
Issue loading.

The Markdown is split into two parts, one for the chart and one for the list,
as that kept the logic separate enough for this usecase.

Reworked code from: ModioAB/nagger!16